re-adjusted the carousel so as to let the images be present in the DOM,  so to easily inplement fade in and fade out image effect
