+++ app/src/components/carousel/carousel.css
@@ -8,7 +8,7 @@
 
 /* Carousel.css */
 
-.carousel {
+/* .carousel {
   width: 100%;
   height: 100vh;
   position: relative;
@@ -36,18 +36,22 @@
 
 .main-container {
   height: 100vh;
-  width: 100%;
+  width: 100vw;
   display: flex;
   align-items: center;
   justify-content: start;
-  padding-left: 5vw;
 }
 
 .preview-box {
-  max-width: 600px;
-  background-color: rgba(255, 255, 255, 0.05);
+  width: 100vw;
+  height: 100vh;
+  background-color: rgba(8, 8, 8, 0.822);
   padding: 2rem;
-  border-radius: 12px;
+  border-radius: 8px;
+  -webkit-border-radius: 8px;
+  -moz-border-radius: 8px;
+  -ms-border-radius: 8px;
+  -o-border-radius: 8px;
 }
 
 .hero-text .head-text {
@@ -71,23 +75,53 @@
   margin-top: 2rem;
 }
 
-.navigation-button {
+.navigation-button{
+  color: white;
   background: linear-gradient(to right, #ff9a27d7, #ff5100dc);
-  padding: 0.75rem 1.5rem;
-  font-size: 1.2rem;
+  padding: 0.5rem 1.5rem;
+  font-size: 1.5rem;
+  font-weight: 500;
   border: none;
   border-radius: 60px;
-  color: white;
   cursor: pointer;
-  white-space: nowrap;
-  transition: background 0.3s ease;
+  margin-block-start: 60px;
+  -webkit-border-radius: 60px;
+  -moz-border-radius: 60px;
+  -ms-border-radius: 60px;
+  -o-border-radius: 60px;
+  position: relative;
+  overflow: hidden;
+  box-shadow: 0 4px 7px rgba(0, 0, 0, 0.2);
+  z-index: 1;
+  text-wrap: none;
+
 }
 
-.navigation-button:hover {
-  background: linear-gradient(to right, #ffc58c, #ff8143);
+.navigation-button::before {
+  content: '';
+  position: absolute;
+  left: 0;
+  bottom: 0;
+  width: 100%;
+  height: 0%;
+   background: linear-gradient(to right, #eedac2, #f19469);
+    opacity: 0.4;
+  z-index: -1;
+  transition: height 0.6s ease;
+  border-radius: 60px;
+  -webkit-transition: height 0.6s ease;
+  -moz-transition: height 0.6s ease;
+  -ms-transition: height 0.6s ease;
+  -o-transition: height 0.6s ease;
 }
 
-/* 🔁 Responsive Queries */
+.navigation-button:hover::before {
+  height: 100%;
+  opacity: 0.4;
+}
+
+
+ 🔁 Responsive Queries 
 
 @media screen and (max-width: 1024px) {
   .main-container {
@@ -167,5 +201,242 @@
     .slide-background {
     background-position: center top;
     background-size: cover; /* Show full image */
+
+    .carousel {
+  width: 100%;
+  height: 100vh;
+  position: relative;
+  overflow: hidden;
+}
+
+/* Each slide stacked */
+.slide {
+  opacity: 0;
+  position: absolute;
+  inset: 0;
+  transition: opacity 1.2s ease;
+  z-index: 0;
+}
+
+.slide.active {
+  opacity: 1;
+  z-index: 1;
+}
+
+/* Background image with zoom */
+.slide-background {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  background-size: cover;
+  background-position: center;
+  background-repeat: no-repeat;
+  transform: scale(1.05);
+  opacity: 0.7;
+  transition: transform 1.2s ease, opacity 1.2s ease;
+  z-index: -1;
+}
+
+.slide.active .slide-background {
+  transform: scale(1);
+  opacity: 1;
+}
+
+/* Container for text + buttons */
+.main-container {
+  height: 100vh;
+  width: 100vw;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.preview-box {
+  width: 100vw;
+  height: 100vh;
+  background-color: rgba(8, 8, 8, 0.65);
+  border-radius: 8px;
+  align-items: center;
+  justify-content: center;
+}
+
+.hero-text {
+  opacity: 0;
+  transform: translateY(20px);
+  transition: all 0.8s ease;
+}
+
+.slide.active .hero-text {
+  opacity: 1;
+  transform: translateY(0);
+  transition-delay: 0.3s; 
+}
+
+.hero-text .head-text {
+  font-size: 3.5rem;
+  line-height: 1.1;
+  font-weight: bold;
+  color: var(--accent-orange);
+}
+
+.description p {
+  font-size: 1.2rem;
+  color: var(--primary-white);
+  margin-top: 1.5rem;
+  line-height: 1.6;
+}
+
+/* navigation buttons */
+.button-container {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 1rem;
+  margin-top: 2rem;
+}
+
+.navigation-button {
+  color: white;
+  background: linear-gradient(to right, #ff9a27d7, #ff5100dc);
+  padding: 0.5rem 1.5rem;
+  font-size: 1.5rem;
+  font-weight: 500;
+  border: none;
+  border-radius: 60px;
+  cursor: pointer;
+  margin-block-start: 60px;
+  position: relative;
+  overflow: hidden;
+  box-shadow: 0 4px 7px rgba(0, 0, 0, 0.2);
+  z-index: 1;
+  text-wrap: none;
+}
+
+.navigation-button::before {
+  content: "";
+  position: absolute;
+  left: 0;
+  bottom: 0;
+  width: 100%;
+  height: 0%;
+  background: linear-gradient(to right, #eedac2, #f19469);
+  opacity: 0.4;
+  z-index: -1;
+  transition: height 0.6s ease;
+  border-radius: 60px;
+}
+
+.navigation-button:hover::before {
+  height: 100%;
+  opacity: 0.4;
+}
+
+@media screen and (max-width: 1024px) {
+  .main-container {
+    justify-content: center;
+  }
+
+  .preview-box {
+    width: 100%;
+    padding: 1.5rem;
+  }
+
+  .hero-text .head-text {
+    font-size: 2.5rem;
+  }
+
+  .description p {
+    font-size: 1rem;
+  }
+}
+
+@media screen and (max-width: 768px) {
+  .main-container {
+    align-items: flex-start;
+    padding-top: 4rem;
+  }
+
+  .hero-text .head-text {
+    font-size: 2rem;
+  }
+
+  .navigation-button {
+    font-size: 1rem;
+    padding: 0.6rem 1.2rem;
+  }
+
+  .slide-background {
+    background-position: center top;
+    background-size: cover;
+  }
+}
+
+/* @media screen and (max-width: 480px) {
+  .main-container {
+    flex-direction: column;
+    align-items: center;
+    padding: 1rem;
+  }
+
+  .preview-box {
+    background: rgba(255, 255, 255, 0.2);
+    padding: 1rem;
+    border-radius: 10px;
+  }
+
+  .hero-text .head-text {
+    font-size: 1.8rem;
+    text-align: center;
+  }
+
+  .description p {
+    font-size: 0.95rem;
+    text-align: center;
+  }
+
+  .button-container {
+    flex-direction: column;
+    align-items: center;
+  }
+
+  .navigation-button {
+    width: 100%;
+    max-width: 300px;
+    font-size: 1rem;
+  }
+
+  .slide-background {
+    background-position: center top;
+    background-size: cover;
+  }
+} */
+
+@media screen and (max-width: 480px) {
+  .preview-box {
+    background-color: rgba(8, 8, 8, 0.65); /* same dark overlay */
+    padding: 1rem;
+    border-radius: 10px;
+  }
+
+  .hero-text .head-text {
+    font-size: 1.8rem;
+    text-align: center;
+  }
+
+  .description p {
+    font-size: 0.95rem;
+    text-align: center;
+  }
+
+  .button-container {
+    flex-direction: column;
+    align-items: center;
+  }
+
+  .navigation-button {
+    width: 100%;
+    max-width: 300px;
+    font-size: 1rem;
   }
 }
